requirements.txt
h5py
numpy
Pillow
pycuda
tqdm
